requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
pydantic>=2.0.0
pydantic[email]
arxiv>=2.1.0
openai>=1.0.0
apscheduler>=3.10.0,<4.0.0
pdfminer.six>=20221105
requests>=2.28.0
pyjwt>=2.8.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
Pillow>=10.0.0
